Move Sign Elements to dedicated service
