+++ lib/Service/AccountService.php
@@ -47,10 +47,8 @@ use OCP\AppFramework\Utility\ITimeFactory;
 use OCP\Files\Config\IMountProviderCollection;
 use OCP\Files\Config\IUserMountCache;
 use OCP\Files\File;
-use OCP\Files\Folder;
 use OCP\Files\IMimeTypeDetector;
 use OCP\Files\IRootFolder;
-use OCP\Files\NotFoundException;
 use OCP\Http\Client\IClientService;
 use OCP\IConfig;
 use OCP\IGroupManager;
@@ -65,9 +63,6 @@ class AccountService {
 	private SignRequest $signRequest;
 	private \OCA\Libresign\Db\File $fileData;
 	private \OCP\Files\File $fileToSign;
-
-	public const ELEMENT_SIGN_WIDTH = 350;
-	public const ELEMENT_SIGN_HEIGHT = 100;
 
 	public function __construct(
 		private IL10N $l10n,
@@ -471,96 +466,6 @@ class AccountService {
 			}
 		}
 		return $content;
-	}
-
-	public function getElementsFromSession(string $sessionId): array {
-		$folder = $this->folderService->getFolder();
-		try {
-			/** @var Folder $signerFolder */
-			$signerFolder = $folder->get($sessionId);
-		} catch (NotFoundException $th) {
-			return [];
-		}
-		$fileList = $signerFolder->getDirectoryListing();
-		$return = [];
-		foreach ($fileList as $fileElement) {
-			list($type, $timestamp) = explode('_', pathinfo($fileElement->getName(), PATHINFO_FILENAME));
-			$return[] = [
-				'type' => $type,
-				'file' => [
-					'url' => $this->urlGenerator->linkToRoute('ocs.libresign.account.getSignatureElementPreview', [
-						'apiVersion' => 'v1',
-						'fileId' => $fileElement->getId(),
-					]),
-					'fileId' => $fileElement->getId(),
-				],
-				'starred' => 0,
-				'createdAt' => (new \DateTime())->setTimestamp((int) $timestamp)->format('Y-m-d H:i:s'),
-			];
-		}
-		return $return;
-	}
-
-	/**
-	 * @return ((int|string)[]|\DateTime|int|string)[][]
-	 *
-	 * @psalm-return list<array{id: int, type: string, file: array{url: string, fileId: int}, uid: string, starred: 0|1, createdAt: \DateTime}>
-	 */
-	public function getUserElements(string $userId): array {
-		$elements = $this->userElementMapper->findMany(['user_id' => $userId]);
-		$return = [];
-		foreach ($elements as $element) {
-			$exists = $this->signatureFileExists($element);
-			if (!$exists) {
-				continue;
-			}
-			$return[] = [
-				'id' => $element->getId(),
-				'type' => $element->getType(),
-				'file' => [
-					'url' => $this->urlGenerator->linkToRoute('core.Preview.getPreviewByFileId', ['fileId' => $element->getFileId(), 'x' => self::ELEMENT_SIGN_WIDTH, 'y' => self::ELEMENT_SIGN_HEIGHT]),
-					'fileId' => $element->getFileId()
-				],
-				'uid' => $element->getUserId(),
-				'starred' => $element->getStarred() ? 1 : 0,
-				'createdAt' => $element->getCreatedAt()
-			];
-		}
-		return $return;
-	}
-
-	private function signatureFileExists(UserElement $userElement): bool {
-		try {
-			$this->folderService->getFolder($userElement->getFileId());
-		} catch (\Exception $e) {
-			$this->userElementMapper->delete($userElement);
-			return false;
-		}
-		return true;
-	}
-
-	/**
-	 * @return ((int|string)[]|\DateTime|int|string)[]
-	 *
-	 * @psalm-return array{id?: int, type?: string, file?: array{url: string, fileId: int}, uid?: string, starred?: 0|1, createdAt?: \DateTime}
-	 */
-	public function getUserElementByElementId(string $userId, $elementId): array {
-		$element = $this->userElementMapper->findOne(['id' => $elementId, 'user_id' => $userId]);
-		$exists = $this->signatureFileExists($element);
-		if (!$exists) {
-			return [];
-		}
-		return [
-			'id' => $element->getId(),
-			'type' => $element->getType(),
-			'file' => [
-				'url' => $this->urlGenerator->linkToRoute('core.Preview.getPreviewByFileId', ['fileId' => $element->getFileId(), 'x' => self::ELEMENT_SIGN_WIDTH, 'y' => self::ELEMENT_SIGN_HEIGHT]),
-				'fileId' => $element->getFileId()
-			],
-			'uid' => $element->getUserId(),
-			'starred' => $element->getStarred() ? 1 : 0,
-			'createdAt' => $element->getCreatedAt()
-		];
 	}
 
 	public function deleteSignatureElement(string $userId, int $elementId): void {
